backend/requirements.txt
fastapi
uvicorn
pydantic
aiohttp
httpx
python-dotenv
elevenlabs
emoji
pydub